drivers: modem: quectel-bg9x: Reset BG95 on initialization
Use the reset pin instead of the pwrkey to reset the BG95
module on initialization. This allows the module to always
power up as opposed to using the pwrkey which will toggle
the module on/off leaving it in an unknown state on boot.

Note the timings were updated according to
Quectel_BG95_Hardware_Design_V1.1.pdf